Split release code in parts
