+++ src/main/scala/package.scala
@@ -1,0 +1,8 @@
+package ohnosequences.sbt
+
+import sbt._
+
+package object nice {
+
+  type DefTask[X] = Def.Initialize[Task[X]]
+}
